update app cart
